added removeExpense method

+++ src/main/scala/engine/Expense.scala
@@ -3,4 +3,4 @@ import Category._
 
 import java.time.LocalDate
 
-case class Expense(id: Int, amount: Int, date: LocalDate, category: Category, description: String=null)
+case class Expense(id: Int, amount: BigDecimal, date: LocalDate, category: Category, description: String=null)
